pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "td"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "langchain-openai",
    "notion-client",
    "python-dotenv",
    "typer",
    "rich",
]

[tool.setuptools]
packages = ["td_agent"]

[project.scripts]
td = "td_agent.cli:app"
